+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## unpublished
 - improved the UI for user executed discrete tasks (pub get, pub upgrade, ...)
 - fixed running sky apps (the `run sky application` command)
+- added a `Settings…` menu item to the Dart package menu
 
 ## 0.4.2
 - fixed an exception from the outline view when viewing empty Dart files